% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/function_finds_repeated_values_inside_sampling_units.R
\name{FindRepetitionsSU}
\alias{FindRepetitionsSU}
\title{FindRepetitionsSU. Function that finds repeated values in sampling units and calculates their frequency}
\usage{
FindRepetitionsSU(data.x, ncol.names.su, ncol.observation, result = "frequency")
}
\arguments{
\item{data.x}{It's a data.frame with observations (numeric or factor) per sampling unit or some category}

\item{ncol.names.su}{Index of the column which has sampling units's names, or cathegories names.}

\item{ncol.observation}{Column's index of the observations's column (factor or numerical).}

\item{result}{If result is equal to "frequency", then FindRepetitionsSU returns a data.frame with 3 columns: $names.su, $observation, $frequency. Last column has the observation frequency per sampling unit. If result is equal to "is.repeated" function adds a new column to the original data set naming with FALSE unique values or observations that have been registered for first time and TRUE repeated observations.}
}
\value{
see parameter result in Arguments
}
\description{
The imput of this function is a data frame of sampling units with their corresponding data or observations. It finds repeated values inside sample units and calculates their frequency.
}
\examples{
# data.set example number 1. 

# It has 5 sampling units, each one with 4 observations of class factor.
col.names.su.1 <- c(rep("M1",4),rep("M2",4),rep("M3",4),rep("M4",4), rep("M5", 4))  # create sample units
simulated.species <- c(rep("sp1",20), rep("sp2",20), rep("sp3",20), rep("sp4",20), rep("sp5",20), rep("sp6",20))  # create a pool of species. From there, obtain a sample of size = 20.
col.factor.observation <- sample(simulated.species, size = 20)
ej1 <- cbind(col.names.su.1, col.factor.observation)
ej1 <-as.data.frame(ej1)
ej1.result.is.repeated <- FindRepetitionsSU(data.x = ej1, ncol.names.su = 1, ncol.observation = 2, result = "is.repeated")
ej1.result.frequencies <- FindRepetitionsSU(data.x = ej1, ncol.names.su = 1, ncol.observation = 2, result = "frequency")

# data.set example number 2. 
# It has 5 different sample units, each one with 4 observations of class numeric.
col.names.su.2 <- c(rep("M1", 1),rep("M2",1),rep("M3",1),rep("M4",1), rep("M5", 1), rep("M1", 2),rep("M2",2),rep("M3",2),rep("M4",2), rep("M5", 2), rep("M1", 1),rep("M2",1),rep("M3",1),rep("M4",1), rep("M5", 1))  # sample units are mixed.
simulated.values <- c(rep(1,20), rep(2,20), rep(3,20), rep(4,20), rep(5,20), rep(6,20))  # create a pool of species. From there, obtain a sample of size = 20.
col.numeric.observation <- sample(simulated.values, size = 20)
ej2 <- cbind(col.names.su.2, col.numeric.observation)
ej2 <-as.data.frame(ej2)
ej2.result.is.repeated <- FindRepetitionsSU(data.x = ej2, ncol.names.su = 1, ncol.observation = 2, result = "is.repeated")
ej2.result.frequencies <- FindRepetitionsSU(data.x = ej2, ncol.names.su = 1, ncol.observation = 2, result = "frequency")
}
\author{
Natalia Mariana Denkiewicz, email = nataliamdenk11@gmail.com
}
