% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/function_calculates_daily_value_for_hourly_data.R
\name{HourlyToDaily}
\alias{HourlyToDaily}
\title{HourlyToDaily. Function that returns a data.frame with accumulated or average daily data from a data frame with hourly data.}
\usage{
HourlyToDaily(data.x, ncol.obs, na.presence = TRUE, value = "mean")
}
\arguments{
\item{data.x}{Data.frame with data recorded every hour.}

\item{ncol.obs}{Index of the column which has hourly observations (numeric).}

\item{na.presence}{If TRUE, variable has NAs in it. If FALSE, it hasn't. Default is TRUE.}

\item{value}{If value = "mean", functions returns monthly mean. If value = "accumulated", it returns the sum. Default is "mean".}
}
\value{
Returns a data.frame with accumulated or avarage daily data.
}
\description{
This function takes a data.frame with hourly data and returns another data.frame with average or accumulated daily data. It needs columns with day, month and year corresponding to each observation, and they can be easily obtained using OrganizeDateColumns function.
}
\author{
Natalia Mariana Denkiewicz, email = nataliamdenk11@gmail.com
}
