% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_from_hourly_data_to_daily_data.R
\name{HourlyToDailyAccum}
\alias{HourlyToDailyAccum}
\title{HourlyToDailyAccum. Function that returns a data.frame with accumulated daily data from a data frame with hourly data.}
\usage{
HourlyToDailyAccum(data.x, ncol.obs, ncol.date = NULL, orders = NULL, ncol.day = NULL, ncol.month = NULL, ncol.year = NULL)
}
\arguments{
\item{data.x}{Data.frame with data registered every hour.}

\item{ncol.obs}{Index of the column which has hourly observations (numeric).}

\item{ncol.date}{Column's index for date's column. Default is NULL (numeric).}

\item{orders}{Only when data.frame has a date column. It's the order of parameters in date, where firts letter represents the component of the date (e.g. day is d). See help(topic = parse_date_time, package = lubridate) (character).}

\item{ncol.day}{Index of the column which has the number of the day. Default is NULL (numeric).}

\item{ncol.month}{Index of the column which has the number of the month. Default is NULL (numeric).}

\item{ncol.year}{Index of the column which has the year. Default is NULL (numeric).}
}
\value{
Returns a data.frame with accumulated daily data.
}
\description{
The imput of this function is a data frame with hourly observations. A column with the day number or the date are needed.
}
\author{
Edgar Federico Rivadeneira, email = federivadeneira@gmail.com
}
