% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/function_calculates_monthly_value_for_daily_data.R
\name{MonthlyValue}
\alias{MonthlyValue}
\title{MonthlyValue. Calculates monthly means or monthly sums from daily data}
\usage{
MonthlyValue(data.x, ncol.interest, value = "mean", ncol.date = NULL, ncol.year = NULL, ncol.month = NULL, na.presence = FALSE).
}
\arguments{
\item{data.x}{It's a data.frame with daily data for some variable.}

\item{ncol.interest}{Column's index of the variable x (numeric).}

\item{value}{If value = "mean", functions returns monthly mean. If value = "accumulated", it returns the sum. Default is "mean".}

\item{ncol.date}{Column's index of the dates's column. Default is NULL.}

\item{ncol.year}{Index of the column which has the year of each x[i, ]. Default is NULL.}

\item{ncol.month}{Index of the column which has the number of the month of each x[i, ]. Default is NULL (numeric).}

\item{na.presence}{If TRUE, variable has NAs in it. If FALSE, it hasn't. Default is FALSE.}
}
\value{
It returns a data.frame with three columns. 1) $month.order: Column with the number of the month (from 1 to 12) corresponding to each value, 2) $year: Column with the year, and 3) $mean.value or $accumulated.value: Column with the monthly value of the observation
}
\description{
Using daily data, calculates monthly values (mean or accumulated value) for all the given years. The imput of this function is a data frame with daily data for some variable. A column with the date or colums with the number of the month and year are needed.
}
\author{
Natalia Mariana Denkiewicz, email = nataliamdenk11@gmail.com
}
