% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_arrange_data_set_in_dmy.R
\name{OrganizeDateColumns}
\alias{OrganizeDateColumns}
\title{OrganizeDateColumns. Function that returns a data.frame with columns of day, month and year.}
\usage{
OrganizeDateColumns(data.x, ncol.date = 1, orders = "mdyHMS", ncol.day = NULL, ncol.month = NULL, ncol.year = NULL)
}
\arguments{
\item{data.x}{Data.frame with observations across time.}

\item{ncol.date}{Index of the column which has date data (numeric). Default is NULL.}

\item{orders}{Only when data.frame has a date column. It's the order of parameters in date. It's formed by the firt letter of each component of the date (e.g. day is d). Day, month, and year are written in lower case. Hour, minute and second are written in capital letters (e.g. hour is H). See help(topic = parse_date_time, package = lubridate) (character).}

\item{ncol.day}{Index of the column which has the number of the day. Default is NULL (numeric).}

\item{ncol.month}{Index of the column which has the number of the month. Default is NULL (numeric).}

\item{ncol.year}{Index of the column which has the year. Default is NULL (numeric).}
}
\value{
Returns a data.frame with columns with the day, month and year corresponding to each observation.
}
\description{
The input of this function is a data frame with observations through time. This function organizes date data and adds (or renames) columns with the day, the month and the year corresponding to each observation.
}
\details{
This function is useful for organizing data before its use in functions that make daily or monthly calculations.
}
\author{
Natalia Mariana Denkiewicz, email = nataliamdenk11@gmail.com
}
