% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/function_from_hourly_data_to_daily_data_accumulated.R
\name{HourlyToDailyAccum}
\alias{HourlyToDailyAccum}
\title{HourlyToDailyAccum. Function that returns a data.frame with accumulated daily data from a data frame with hourly data.}
\usage{
HourlyToDailyAccum(data.x, ncol.obs)
}
\arguments{
\item{data.x}{Data.frame with data recorded every hour.}

\item{ncol.obs}{Index of the column which has hourly observations (numeric).}
}
\value{
Returns a data.frame with accumulated daily data.
}
\description{
This function takes a data.frame with hourly data and returns another data.frame with accumulated daily data. It needs columns with day, month and year corresponding to each observation, and they can be easily obtained using OrganizeDateColumns function.
}
\author{
Edgar Federico Rivadeneira, email = federivadeneira@gmail.com
}
